DO NOT MERGE! zvkg: Adding vgmult.vv and vghash.vv
!!!
This patch introduces made up instructions with an instruction
encoding that conflicts with the RISC-V encoding space.
Do not use this patch, if you don't know what that means.
!!!

This patch adds two additional instructions to Zvkg:
* vgmult.vv vd, vs2         # Y[i+1] = (Y[i] ⋅ H)
* vghash.vv vd, vs2, vs1    # Y[i+1] = ((Y[i] ^ X[i]) ⋅ H

These instructions are not part of any specification.

The already implemented vghmac.vv instruction remains the same:
* vghmac.vv vd, vs2, vs1    # Y[i+1] = (Y[i] ⋅ H) ^ X[i+1]

Signed-off-by: Christoph Müllner <christoph.muellner@vrull.eu>