DO NOT COMMIT: Fixup for AES test
The out of tree series from Richard Henerson,
which is required as a dependency for the RISC-V vector crypto
extension support series, introduces tests for AES using
the AES instructions.

The compilation of this test is gated by the compiler accepting
"-mrv64g_zk". As this is not a valid compiler flag (valid would
be "-march=rv64g_zk"), the AES test will never be executed.

This fixes this by dropping the gate and using the RISC-V-specific
".insn" directives to build up the AES instructions.

With this change applied, there is not depenencies to the cross
toolchain anymore and the test will always be executed.

Signed-off-by: Christoph Müllner <christoph.muellner@vrull.eu>